requirements.txt
Authlib==1.0.1
cryptography==3.3.2
Flask==2.1.2
Flask-Async==0.11.dev0-20140215
Flask-Authlib==1.6.0
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
jwt==1.3.1
PyJWT==2.4.0
Werkzeug==2.1.2
Flask-Admin==1.6.0
pytest==7.1.2
flake8==4.0
redis==4.3.4
Flask-JWT-Extended==4.4.2
SQLAlchemy==1.4.39